[quant][core][improvement] Improved efficiency of fill_quantized by using broadcasting
Summary:
Previously, fill_quantized was inefficient as it involved creating a
tensor of constants, quantizing all of those constants individaully, and
then assigning those constants to the output tensor.

Rather than doing that, we do this without broadcasting, eliminating the
need to create a quantized tensor of constants.

Test Plan:
TBA

ghstack-source-id: e8ae5b7c9fae0269da7f24079e49e042c1cd0b29
Pull Request resolved: https://github.com/pytorch/pytorch/pull/78621